added report app

+++ app/triage/serializers.py
@@ -7,7 +7,7 @@ class SecurityEventSerializer(serializers.ModelSerializer):
     asset_vendor     = serializers.CharField(source='asset.vendor_oui', read_only=True)
     asset_type       = serializers.CharField(source='asset.asset_type', read_only=True)
     asset_rssi       = serializers.IntegerField(source='asset.smoothed_rssi', read_only=True)
-    asset_radius     = serializers.FloatField(source='asset.estimated_radius_meters', read_only=True)
+    asset_radius     = serializers.IntegerField(source='asset.estimated_radius_meters', read_only=True)
     asset_channel    = serializers.IntegerField(source='asset.operating_channel', read_only=True)
     asset_first_seen = serializers.DateTimeField(source='asset.first_seen', read_only=True)
